+++ CHANGES.md
@@ -9,6 +9,7 @@ Changes in 6.0
 - JS update jQuery to v2.2.4 in jquery.js
 - JS MarkDown use marked instead of showdown (15KB smaller) in assets/js/marked/, plugins.min.js, warehouse.js, warehouse_wkhtmltopdf.js & PDF.php
 - JS replace jqPlot with Chart.js in assets/js/Chart.js/, stylesheet.css & Charts.fnc.php
+- Do NOT remove Grade Level once has Student enrolled in GradeLevels.php
 
 Changes in 5.9.2
 ----------------